requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
python-socketio>=5.11
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20
anthropic>=0.40
python-dotenv>=1.0
jinja2>=3.1
python-multipart>=0.0.9
httpx>=0.27
pydantic>=2.0
